% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ensembl_query_func.R
\name{create_ensembl_query_func}
\alias{create_ensembl_query_func}
\title{Create a function to query genes}
\usage{
create_ensembl_query_func(ensembl, full_genes_only = TRUE)
}
\arguments{
\item{ensembl}{Ensembl gene annotations as a GenomicRanges object.}

\item{full_genes_only}{If TRUE, filter by \code{type=="gene"}}
}
\value{
Function with three arguments, \code{chr}, \code{start},
    and \code{end}, which returns a data frame with the genes
    overlapping that region, with \code{start} and \code{end} being
    in Mbp. The output should contain at least the columns
    \code{Name}, \code{chr}, \code{start}, and \code{stop}, the
    latter two being positions in Mbp.
}
\description{
Create a function that will query genes from a GenomicRanges object with
ensembl gene annotations and return a data frame with gene
information for a selected region.
}
\details{
Note that this function assumes that the database has
    \code{start} and \code{end} fields that are in basepairs, but
    the selection uses positions in Mbp, and the output data frame
    should have \code{start} and \code{stop} columns in Mbp.
}
\examples{
# small version of ensembl data
ensembl <- readRDS(system.file("extdata", "ensembl_small.rds", package="qtl2bioc"))

# create query function, pulling only full genes
query_ensembl <- create_ensembl_query_func(ensembl)

# genes on chr 2 overlapping (96.5 - 97.0)
genes <- query_ensembl("2", 96.5, 97.0)
}
