% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_ensembl.R
\name{grab_ensembl}
\alias{grab_ensembl}
\title{Grab ensembl gene annotations}
\usage{
grab_ensembl(organism = "mus musculus", set = "AH51040")
}
\arguments{
\item{organism}{Character string with name of organism}

\item{set}{Character string with name of dataset}
}
\value{
A GenomicRanges object with the gene annotation information.
}
\description{
Grab ensembl gene annotations from the Bioconductor AnnotationHub.
}
\examples{
\dontrun{
ensembl <- grab_ensembl()
}
}
