% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl2bioc-package.R
\docType{package}
\name{qtl2bioc-package}
\alias{qtl2bioc}
\alias{qtl2bioc-package}
\title{qtl2bioc: Connect to Bioconductor for QTL Experiments}
\description{
Functions to connect to genome databases via Bioconductor, such as to obtain gene annotation information. Part of R/qtl2, a reimplementation of the R/qtl package to better handle high-dimensional data and complex cross designs.
}
\section{Vignettes}{

\itemize{
\item \href{https://kbroman.org/qtl2/assets/vignettes/user_guide.html}{user guide}
\item \href{https://kbroman.org/qtl2/assets/vignettes/input_files.html}{input file formats}
\item \href{https://kbroman.org/qtl2/pages/prep_do_data.html}{preparing DO mouse data for R/qtl2}
\item \href{https://kbroman.org/qtl2/assets/vignettes/rqtl_diff.html}{differences between R/qtl and R/qtl2}
\item \href{https://kbroman.org/qtl2/assets/vignettes/developer_guide.html}{developer guide}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://kbroman.org/qtl2/}
  \item \url{https://github.com/rqtl/qtl2bioc}
  \item Report bugs at \url{https://github.com/rqtl/qtl2bioc/issues}
}

}
\author{
\strong{Maintainer}: Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})

Authors:
\itemize{
  \item Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})
}

}
\keyword{internal}
